requirements.txt
flask>=2.3.0
flask-jwt-extended>=4.5.0
flask-cors>=4.0.0
flask-bcrypt>=1.0.0
pymongo>=4.5.0
python-dotenv>=1.0.0
gunicorn>=21.0.0
